pluginscontainer: Use PluginProvider to access plugins of parent
Instead of copying the parent's plugins into the child's plugins
container, we now use a PluginProvider to access the parent's plugins.
This better handles the case where the parent (e.g. Workspace) is
refreshed but the child is not. Any closeable plugins (like maven
repos) in the parent would linger in the child's plugins container.

Since the child no longer contains the parent's plugin objects,
it always has the correct view onto the parent's plugins.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>