[baseline] Change in builder cased baseline errors no longer properly reported
The recent builder changes passed a Processor to the marker handling instead
of a Project. This was the proper current API. However, the base class
had a bridge method to the original method that took a Project. 

The BaselineErrorHandler therefore implemented the Project method although
it did not require a Project. The bridge method therefore failed silently.

I've deprecated the old method and added a warning to the log if the
old method is still used. The only implementation of it is in the
BaselineErrorDetailsHandler and it has been properly updated.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>