[markers] Markers lost their position
Originally the error/warning Location was set on the place where
the problem was detected. However, we switched to just
set the header & clause. The Default Error Handler then 
would lookup the header and clause and set the line, start, and
and end attributes.

However, since we delayed the reporting, we gave a processor.
This was a bare processor, so the findHeader had neither
the proper file, nor an inheritance chain, to find the
header clause combination. So all errors ended up at the
top line.

This patch makes the temporary processor that holds
the messages inherit from project.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>