plugins: Reduce to a single distinct call to remove duplicate plugins
We move the distinct call to the "exit" point rather than calling
at each processor in the hierarchy when assembling the stream.

The issue with sub bundles was that the ProjectBuilder for a subbundle
has the main project's ProjectBuilder as its parent so the Builder
plugins are added at two levels of the hierarchy. But is also possible
for plugins to be duplicates on the hierarchy, so a distinct call is
needed to provide set semantics (no duplicates).

Signed-off-by: BJ Hargrave <bj@hargrave.dev>