plugins: Provide a hierarchical ordering for plugins
ParentPluginProvider led to an ordering problem where a plugin from the
parent would be selected over a plugin from the child. This issue
had been present before but ParentPluginProvider did not help.

So we remove ParentPluginProvider and define a PluginsSpliterator
which defines an ordering where the child's plugins are before the
parent's plugins. This also allows us to simplify the stream code
and avoid using flatMaps over a single item stream and using concat
to combine two streams. This is all handled in the spliterator
implementation by walking multiple spliterators in an order.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>