launcher: Refactor to call calculatedProperties later
We no longer call calculatedProperties during updateFromProject since
that is too soon. We now call calculatedProperties from getRunProperties
so it is done just as it is needed.

For ProjectLauncher.calculatedProperties, we no longer set a number of
values since they are already handled in
ProjectLauncherImpl.getConstants as launcher properties which are
then set as framework properties by Launcher.createFramework. So we
then needed to modify RemoteProjectLauncherPlugin to override
calculatedProperties to set the framework properties.

Fixes https://github.com/bndtools/bnd/issues/5061

Signed-off-by: BJ Hargrave <bj@hargrave.dev>