baseline: Fix computation of suggestedIfProviders
Since we know the interface has a MAJOR bump, any downgrade due to
changing to be a ProviderType will be a MINOR bump. Using the delta
computed by ignoring the interfaces with MAJOR changes, could result
in an UNCHANGED delta which previously resulted in no bump to the older
version. So we now indicate a MINOR bump is requested.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>