bundle annotations: Support replacement of enum types
Using enum types in the CLASS retention bundle annotations is causing
issues for those using the annotations. Various tools, such as javadoc,
attempt to reify the elements in the annotations and since the
osgi.annotation jar is generally a scope=provided dependency, the enum
types are not available to downstream users of the jars using the
OSGi annotations and so tools generates an annoying warning.

See https://github.com/quarkusio/quarkus/issues/19970 and
https://github.com/eclipse/microprofile-config/issues/716.

We support the use of enum values or string values as the annotation
element value through existing conversion support. The OSGi change in
https://github.com/osgi/osgi/pull/404 will move from using enum values
to use string values which are equivalent to the former enum value
names.

We seamlessly handle old and new annotations using the old enum values
or the new string values. Prior to this fix, Bnd already handled the
change through the Converter which converts the string value to an
internal enum value. This change avoids the internal enum type and
processes the string value or string name of the enum value when
processing older versions of the OSGi annotations.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>