[resolve] Repository validation
Updating an older repository set can be painful. We already had the `bnd mbr` command to do some validation and updating of the mvn files. However, these commands lacked an overview.

This PR updates the `bnd resolve validate` command. It used to require a OBR index, and it can still accept it. However, without the index, it will assume that it is in a workspace. It will then resolve all files in the workspace and all configured repositories stand alone. 

Added the following options:

-x Print a requirement -> set resource for missing resources

-u print unused resources in the repository. That is, no workspace resource requires it.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>