reproducible: Update -reproducible to specify the timestamp
To properly support project.build.outputTimestamp for the new jar
goal in bnd-maven-plugin, we need to support allowing the user to
specify the desired timestamp value. So -reproducible can now
accept the same values that maven-jar-plugin supports for
project.build.outputTimestamp.

This required a small change to Builder so that getBuildJar() sets
the reproducible and compression values on the jar.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>