[http client] Limit the # of concurrent connections
This change fixes  #5035

Sometimes a host limits the number of connections
because it sees too many connections from one client
as an attack.

This patch enables a `maxConcurrentConnections` field in the Maven XML or in the ServerDTO. If this is 0, no limit is applied. Otherwise the connection to that host will be limited to that number. Different hosts can have different numbers.

The implementation is unsophisticated. The client uses
the URL Connection Handler to get this limit from the
ServerDTO. If not zero, he URL Connection Handler is then used
as a key in a map to a Semaphore that is initialized with the
URL handler maxConcurrentConnections value.

Each connection to that host will then first acquire a
token and will return it at the end.

To simplify the code, a dummy semaphore is returned for
non limited cases.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>