[search] Event loop exception using search dialog #4957
A few years ago I replaced all image handling with one
single handling in Icons. 

I reasoned that since this was a static cache, I would
not run into any issues from a life cycle perspective. 
However, this issue indicates that it is still possible
to get a disposed image.

Since this image comes from the Icons handler, I am
not checking if the image has been disposed when
I find it in the cache. If so, it is recreated.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>