[libg] Restore old package
When we moved to 6.1 (not a major release) we lost
the package aQute.lib.exceptions because we also
made it available via aQute.bnd.util.

This would've been a nice use case for my shading
code but until we've got something like that we cannot
just remove a package in a minor release.

This is causing so much pain in all my builds that
I am restoring it, unfortunately now having a 
duplicate. 

I am ok to deprecate one of them but this is
a giant pita.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>