projecttracker: Update for reproducibility and avoid CME
The project tracker did not have an ordering in how projects were
processed and returned. So we sort the dir names and also return the
projects in this sorted order to provide consistent actions
across systems.

We also avoid leaking a live copy of the values through the notifier
by making a copy of the values.

We also remove the use of Map.compute. All of these should ameliorate
the CME (ConcurrentModificationException) in #5121.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>