library: Don't use Workspace repo for Workspace
When processing a -library instruction in the Workspace object, we must
not use the Workspace repo to find resources since we are in the middle
of initializing the Workspace upon which all Projects depends and
inherit from. Attempting to locate Workspace library references in
workspace Projects is a form of a cycle.

So when processing -library for the Workspace, we avoid the Workspace
repo.

There are still issues with Projects getting -library info from other
Projects due to ordering issues in initialization of Projects. Some
Projects will necessarily initialize before other Projects and thus may
not locate -library info in yet to be initialized Projects.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>