bndpomrepo: Fix initialization problems and move poll work
We fix the initialization issues in BndPomRepository. A memoizer is
used to make sure there is only one call to the preparation code
and the memoized value is the promise of the eventual completion
of preparation. So both prepare and init will ensure that preparation
has commenced and init will then wait for it to complete.

We also move the poll method call to the normal executor.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>