decoration: Update add literals to add beginning literals to front
We need to be able to decorate and add some literals to the front of
the result. The current code always added at the end but this would
mean that adding an exclusion glob would not be useful since they
must come before a matching inclusion glob. So we change to add any
unused literals at the start of the decoration to the start of the
result. Any unused literals that come after a used pattern or
non-literal pattern. We otherwise preserve the order of the
unused literal as they are added to either the beginning or end of the
result.

We also do not include negated and unused literals in the result since
their literal value is not negated.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>