[doc] Added custom search bar
Most searches end up in the jungle since bnd
is not the most popular tool. Added a search bar
that limits the search to *.bndtools.org & 
bnd.discourse.group.

Seems to be quite effective.

Also removed the ugly red blob for the release 
and replaced the color with a menu indicator. 

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>