[analyzer] Internal attributes not well defined
It was observed that Analyzer decided to not
learn a package if there was already an attribute
set on that package in the manifest.

However, over time lots and lots of 'internal'
attributes entered the fray.

This patch introduces a AttributeClasses enum that
provides the definition of the different attribute
classes and implements a predicate to test this.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>