container: Add new constructor which takes bsn
The Project knows the bsn, but there are was no constructor to use.
So we add a new constructor and use it to capture the actual bsn.

For other cases, we also add a method to compute the bsn from a file
by using the file name and removing the trailing ".jar", if any.

Fixes https://github.com/bndtools/bnd/issues/5090

Signed-off-by: BJ Hargrave <bj@hargrave.dev>