workspace: Add methods for write locking downgrading to read locking
These method take a callable and a function. The callable is called
while holding the write lock. The lock is downgraded to the read lock
and then the function is called with the result of the callable. The
result of the function is finally returned after releasing the read
lock.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>