bnd-baseline-maven: Place baseline info in the exception
When baselining fails in m2e, you only get an exception that it failed
but you have no useful information since that we only sent to the slf4j
log which you don't see in Eclipse. So we change the plugin to behave
like the gradle baseline task. It now writes a baseline report to a
file in the output directory and, if the baseline fails, the contents
of the report are made part of the failure exception.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>