bnd-maven-plugin: Set alternate source for classpath directories
In command line maven, maven always supplies artifact files which are
the built jar. Either in m2 or in a project's target folder.

However, in m2e, probably since the jar goal is not normally executed,
the artifact file for another project the reactor is for the project's
target/classes folder. This causes jar name matching in
-includeresource to fail since it matched in the file name and all the
file names for these project artifacts is "classes".

So we change the plugin to supply the directory classpath entries with
a file name which is synthesized from the artifact information. This
will then be matched to -includeresource references in m2e as well as
command line maven.

Fixes https://github.com/bndtools/bnd/issues/5081

Signed-off-by: BJ Hargrave <bj@hargrave.dev>