library: Do not include Projects in search space for -library
We can't use ResourceRepositoryStrategy.ALL since we are
initializing the Workspace or a Project and looking in
the WorkspaceRepositoryDynamic requires us to initialize
AND prepare ALL projects so the
WorkspaceRepositoryDynamic can index their built files
for capabilities. This would be a circularity problem. We
need to be able to initialize the Workspace and all
Projects before we prepare any Project. Initializing the
Workspace or a Project must not result in preparing any
Project.

Fixes https://github.com/bndtools/bnd/issues/5190

Signed-off-by: BJ Hargrave <bj@hargrave.dev>