bndtools container: Allow user to exclude project and have only jar
This is useful for version=snapshot or version=latest when the
project's built jar contains transformed classes and we want the
Eclipse compiler to use the jar's classes and not the project's
untransformed classes which lead to compilation errors.

By adding `ide=jar-only` to the -buildpath entry, Bndtools
will include the built jar and not include the project in the
class path container.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>