bndtools container: Rename ide to source
To be more general, we reframe the attribute to `source` with the
default value of `project`. This is the prior behavior.

So saying `source=none` on a -buildpath entry, Bndtools
will include the built jar and not include the project in the
class path container.

In the future, other values of `source`, such as a file, could be
specified. But for now, the value `project` includes the project in
the classpath. All other values do not include the project.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>