bndtools container: Use only jar for classes when version=latest
When using version=snapshot or version=latest, we want Eclipse to use
the class files in the project's built jar and not the project's bin
folder. The project's built jar may contain transformed classes and
we want the Eclipse compiler to use the jar's classes and not the
project's untransformed classes which can lead to compilation errors.

When including the project entry in the class path container, Bndtools
will make the project's classes non-accessible which will make Eclipse
use the classes from the project's built jar. Source access via F3
still works.

This replaces the previous attempt of adding a `source` attribute which
the user must set.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>