bndtools container: Only use project for runtime when version=project
When using version=snapshot or version=latest, we want Eclipse to use
the class files in the project's built jar and not the project's bin
folder. The project's built jar may contain transformed classes and
we want the execution (including testing) to use the jar's classes and
not the project's untransformed classes which can lead to
NoClassDefFoundErrors.

So for the runtime class path, we exclude the project bin folders when
the project is not `version=project`.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>