project: Support a "project" attribute to name the bsn source project
Normally, Bnd expects the bsn of a bundle built in the workspace to be
prefixed by the project name. This convention allows a quick lookup
from a bsn to find the project which build the bundle with the bsn.

However, sometimes you need to refactor code, or when transforming,
use a completely different bsn for one of the sub bundles. The bundle
building works fine. But if another project refers to the bsn which
is not prefixed by the project name, Bnd cannot find the bundle and
fails.

Bnd could just exhaustively search the workspace but this would be cost
prohibitive in very large workspaces such as Open Liberty.

So this change introduces a "project" attribute which can be used on
a buildpath/testpath entry to name the project in which to look for the
bsn. So this does not affect the normal project-name=bsn case but allows
for bsn refactoring in a build at the cost of specifying the "project"
attribute. Fortunately, decoration can be used to centralize the
mappings of bsn to project name.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>