gradle: make generated manifest available as effective manifest
When the BundleTaskExtension complete building a bundle, it now takes
the generated manifest of the bundle and makes it available from the
jar task's manifest.effectiveManifest property. This is similar in
behavior to the old gradle "osgi" plugin. So execution of later tasks
can access the information in the generated manifest.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>