bndtools container: Address regression in Open Type Hierarch\y
The JDT type hierarchy search does not work properly when a
source folder is not-accessible following by a binary jar which is.
This caused types in the jar to be missing from the hierarchy result.

So we add back the `source=project` -buildpath clause option. The
value `project` is the default. So saying `source=none` on a -buildpath
entry, will cause Bndtools to mark the source project as discouraged
to force the compiler to use the classes in the generated jar.

Fixes https://github.com/bndtools/bnd/issues/5250

Signed-off-by: BJ Hargrave <bj@hargrave.dev>