[decoration] Backward compatibility
The current decoration rewrite was not backward
compatible since it added literals for the `-runbundles`.

After some discussion we decided to add `++` for decorations
that add literals and a single `+` for the 6.2 behavior.

Since we use the decoration in our own build, this
patch first adds `++` syntax but both the single and double
plus will add literals.

After this works, we need to add test for the single `+` 
(not adding literals) and replace our own use of decoration
to use double plusses where applicable.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>