[decoration] Backward compatibility
The decoration changes to always include literals was not backward
compatible since it added literals for the `-runbundles`.

After some discussion we decided to add `++` for decorations
that always add literals and a single `+` for the 6.2 behavior
of not adding literals.

Since we use decorations requiring adding literals in our own build,
this changes adds support for `++` decoration while `+` decoration
still adds literals.

After this build is complete, we will following with changing our
build to use `++` where necessary and also change the `+` behavior
back to not adding literals.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>
Signed-off-by: BJ Hargrave <bj@hargrave.dev>