jarsigner: Fix 2 bugs in the jarsigner support
First, the support did not consult the JAVA_HOME env var to locate the
jarsigner tool. So this could be a different version than the desired
one.

Second, the support did not handle EC key signing as the *.EC files were
not copied in the result jar which breaks the signing.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>