gradle: Handle Providers in manifest attribute values
BundleTaskExtension supports using the Jar manifest property to
set attributes in the manifest main attribute section. However
it did not properly handle the use of a Provider to supply the value.
This fix is now in line with how the Gradle Manifest object handles
converting attribute values to a String.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>