gradle: Make default Bundle-SymbolicName and Bundle-Version inputs
If Bundle-SymbolicName and/or Bundle-Version are not specified in the
bnd instructions, then BundleTaskExtension will set default values
based upon the Jar task's archiveBaseName, archiveClassifier, and
archiveVersion properties. However, these Jar task properties are not
considered inputs to the Jar task. They are just considered as
parts of the archiveFile property which is an output property.

Setting the defaults as input properties means that Gradle will consider
them as part of the build cache key. This is important since the values
can end up in the build jar's manifest.

Fixes https://github.com/bndtools/bnd/issues/5279

Signed-off-by: BJ Hargrave <bj@hargrave.dev>