[cached resolve] The test failed on Windows
I think this was caused by faster window machines.
The reason for the operation was USE_CACHE which
can only happen if the bndrun file did not pick up
the time of the included file.

I've added a 100 ms delay before taking a time,
this 100 ms should be more than the resolution of
the file system.

So setting the time on the included file was not
seen as later than the project file.

I also added checks to see what the cache reason is
before we call getRunBundles().

Lets see what happens

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>