[mac m1] Added a bndrun file for the m1
The lowest version SWT for the aarch64 architecture
had a requirement on a later swt than we actually
use. The platform swt jar was available from
Maven central but unfortunately the fragment host
swt jar not. 

I temporarily put the swt in a jars directory 
and patched the bndrun file from the x86 version
since this does not support resolving.

It would be nice if we could get this later swt & the 
swt aarch64 from the same place as all eclipse files
so I leave this open for a discussion how to do this.

Unfortunately, I ran into a deadlock when I ran the
code. The builder gets a read lock on the workspace
but calls refresh that tries to get a write lock.

I've added a check for this and report an error if
a read lock holder tries to get the write lock. 
However, I'd like to go to an Update/Read/Write
lock as in https://github.com/npgall/concurrent-locks

Also to discuss.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>