[jvm crash in swt] It is still happening
As a try, I've increased the coalesce time
and then also tried to rearrange some of the overridden
methods. Instead of calling the super last, moved it
to first.

Bit desperate ... it is pretty ok now, it happens
to me once every month. But very frustrating that
we cannot pin this down.

If anybody has a full crash report, appreciated.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>