[clause manager] A clause manager for Processor
See https://github.com/bndtools/bnd/issues/5363

This is the first part of #5363

This part adds a clause manager that allows
code to set arbitrary clauses instead of doing
lots of string processing. If a property is gotten
that was managed with addClauses, this property
will be consolidated. This goes through the
plugins: see ConsolidatedParameters.

The code intents to be completely backward 
compatible. 



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>