#5336 Need -include loop detection
The problem was that there was no test cases. There
actually was a loop detection. I already was incredibly
surprised because I am pretty sure I wrote this code
and forgetting to check cycles is even beneath me.However,
I'd forgotten to add a return after the error message ...

So the code continued forever to include the same
file anyway although dutifully reporting it.

Fixes #5336

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>