[m2e] Expand bundle version ranges for M2E imports
In recent versions (2022-06 and later) of Eclipse the version of the M2E maven runtime bundle has stopped matching the M2E version and now uses a higher version related to the maven it packages. We need to be able to use this if M2E support is to keep working.

Fixes #5417

Signed-off-by: Tim Ward <timothyjward@apache.org>