[m2e] Remove references to internal packages and avoid static methods
The latest m2e updates have moved methods in internal packages. This breaks m2e support in Bndtools where we were going around the official API. This is primarily an attempt to remove the usage of internal types in an effort to remain compatible between versions. Furthermore we need to coexist in scenarios where:

* A maven project change listener whiteboard is built in to M2E 2+ and replaces the one we created
* Services are used for M2E objects such as IMaven - access these in a consistent way
* Public lookup methods now exist, but we must still use reflection while 1.x is supported

A good deal of cleanup will be possible once we have a base of 2.0, but for now this is a good next step.

Known issues: The bndrun "Browse Repos" section populates, but then disappears again after about a second. Resolving is possible.

Signed-off-by: Tim Ward <timothyjward@apache.org>