[m2e] Avoid runtime errors due to API changes between 1.x and 2.x
Several types in M2E have been updated to use Lists rather than arrays. We have to implement the extra interface methods (where needed) and use reflective calls to avoid linking to a method which may not exist. All of this code is ugly, and should be removed once we can use a base of M2E 2.x.

Signed-off-by: Tim Ward <timothyjward@apache.org>