[m2e] Ensure a linkable method exists in M2E 2.x
When running in M2E 2.x the  delegates to the array version of the mavenProjectChanged method, however as this is declared by the 1.x interface the JVM fails to link the method call. We therefore re-declare the method in the Abstract class to give the linker something to link to.

Signed-off-by: Tim Ward <timothyjward@apache.org>
Signed-off-by: BJ Hargrave <bj@hargrave.dev>