maven: Support resolving bndrun files to an output folder
Writing bndrun files in place after resolving is awkward since this
results in modifying files under source control.

So we add the outputBndrunDir configuration parameter to
bnd-resolver-maven-plugin so you can specify the output directory
to write the output bndrun files.

We add the bndrunDir configuration parameter to
bnd-export-maven-plugin and bnd-testing-maven-plugin so they can use
the output bndrun files from a bnd-resolver-maven-plugin execution.
For completeness, we also add the bndrunDir configuration parameter to
bnd-resolver-maven-plugin.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>