maven: Refactor build for maven 4
Maven 4 will use the .mvn folder (like gradle uses the settings.gradle
file) to find the base of the project hierarchy. Since we have .mvn in
the base of the repo, we need to add a reactor pom to the base of the
repo to be the base project.

We also rename the maven folder to maven-plugins to clearly indicate
the project types in the folder.

Signed-off-by: BJ Hargrave <bj@hargrave.dev>