#5462 Limit the Launcher to set startlevels to only the bundles in scope
During bnd 6.2.0 the behavior of the launcher was changed to set start levels on all bundles installed, even if they were not listed in the run bundles scope. This was recently discovered to wreak havoc in systems that use an agent to install additional bundles. I.e. the core set of run bundles were used to create an executable JAR that provided an environment that allowed the agent to install the bundles needed to satisfy the constraints of a remote management system. This management system also provided start levels. The change caused the launcher to set all bundles the agent introduced to the default level. Since the agent used higher numbers, any required ordering was lost.

This change introduces the idea of _narrow_ and _all_ scope. It provides a `-launcher` instruction that can set the `manage` option. This options is either `all` (default), `narrow`, and `none`. This `manage` scope is not limited to start levels in the future. It indicates that the launcher must be aware of additional bundles that are not in its run bundles.

The Project Launcher uses this instruction to set `launch.startlevel.default` to either 0 (none), a negative value (narrow), or a positive value for all. This is then processed appropriately in the Launcher.

The default for the manage instruction is set in the current 7.0.0 defaults.bnd

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>