Make BND compatible with SLF4J-2
SLF4J-2 is a binary compatible drop-in replacement for SLF4J-1 and
therefore artifacts compiled against the SLF4J-1 API can use SLF4J-2 as
well.
To make BND compatible with SLF4J-2 it is sufficient to widen the
'Import-Package' version-range for org.slf4j packages to "[1.x,3)".
This change allows to use BND in runtimes where either SLF4J-1 or
SLF4J-2 is present.

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>