Drop back to Java 11 as a base for the bnd launcher
The bnd launcher code is deliberately minimal as it is used in any exported runnable application. The move of bnd to Java 17 has suddenly required that everyone using the bnd launcher uses Java 17, even if there is no other part of their runtime that needs Java 17. This commit rolls the launcher back to a Java 11 base, while leaving the rest of bnd at Java 17. This allows applications to build and export using bnd 7.x and Java 17, but to still target Java 11 as a base runtime.

Signed-off-by: Tim Ward <timothyjward@apache.org>