#5407 Decision to include project bin directory
Bnd was changed a couple of years ago to have a default expand when there were no instructions like -includeresource, Private-Package, etc. The implementation decided to put the whole output directory of the project directly in the to be build JAR. This meant that the normal processing did not happen. You had to use -exportcontents to export packages.

This PR changes the following

- It introduces a magic keyword for the `from:` directive. If this is `project`, it will search the class path entry that is the project's output directory and will include only from there. Therefore, doing a `-includepackage *;from:=project` will include the _whole_ project output directory.
- Created a set of EXPAND_HEADERS in Constants. The -includresource was not include.
- Updated the Project builder to set the `-includepackage *;from:=project` instruction in the builder when none of the EXPAND_HEADERS was set

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>