#5060 No automatic build when build.bnd changes
I could not get Eclipse to recognize the dependency between the cnf project changes and the other projects. So I used the new notifier mechanism to get changes in the workspace. When the repos are reported to be ready (and it is not the first time) a new full build is started.

Notice that whenever the registration is done, the notifier will send init/repos. That is why the first time it is ignored.

---
 Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>

Signed-off-by: Peter Kriens <Peter.Kriens@aQute.biz>