Fix for issue #5239
Work around to allow NewBndProjectWizardPageOne to function as subclass of
NewJavaProjectWizardPageOne in Eclipse 2022-12 release.   

As described in this comment:
https://github.com/bndtools/bnd/issues/5239#issuecomment-1399331939
starting in Eclipse 2022-12 going forward a module section/control
was added to NewJavaProjectWizardPageOne class.  This UI is also used
in the validation done by NewJavaProjectWizardPageOne.  In the
subclass implementation of NewBndProjectWizardPageOne, the createControl
method *does* not initialize the module section, as it's not needed for
Bnd project creation.   Since the NJPWPO validation depends upon
the module section/control initialization, the NewBndProjectWizardPageOne
does not function properly in Eclipse 2022-12 and going forward.

This work around overrides methods and introduces initialization 
(but not presentation) of the module section in superclass into 
NBndPWPO class.

Signed-off-by: Scott Lewis <scottslewis@gmail.com>